+++ R/MetricDictionary.R
@@ -1,0 +1,17 @@
+#' Metric Dictionary
+#'
+#' This data frame contains metrics from all tests.
+#'
+#' @format A data frame with 484 rows and 7 variables:
+#' \describe{
+#'   \item{canonicalTestTypeId}{The unique id of the canonical test type (character).}
+#'   \item{testTypeName}{The given / common name of the test type (character).}
+#'   \item{id}{CamelCase formatted name of the metric (character).}
+#'   \item{label}{The metric label found in app and cloud UI (character).}
+#'   \item{label_unit}{The metric name returned from the API (character).}
+#'   \item{units}{Metric unit of measure (character).}
+#'   \item{description}{A verbose description of the metric (character).}
+#' }
+#' @source Generated for demonstration purposes.
+"MetricDictionary"
+