+++ R/get_tests_ath.R
@@ -223,7 +223,7 @@ get_tests_ath <-
         trialMetrics <- df[6:ncol(df)]
 
         # Clean Trial Metric Names
-        trialMetrics <- janitor::clean_names(trialMetrics)
+        trialMetrics <- replace_headers(df = trialMetrics, typeId = "all")
 
         ### Test Type Data -----
         # Index Test Type Data